# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "User Registeration",
                "description": "Register User Here",
                "operationId": "Register",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation",
                                    "whatsappNumber",
                                    "countryCode",
                                    "user_name",
                                    "first_name",
                                    "last_name",
                                    "country"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "user_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "JohnD"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "whatsapp_number": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "0502342345"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "GH"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Ghana"
                                    },
                                    "town_city": {
                                        "type": "string",
                                        "example": "Accra"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/verifyOtp": {
            "post": {
                "tags": [
                    "VerifyOtp"
                ],
                "summary": "Verify OTP",
                "description": "Verify OTP",
                "operationId": "VerifyOtp",
                "requestBody": {
                    "description": "Verify user otp",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "otp",
                                    "whatsappNumber",
                                    "countryCode"
                                ],
                                "properties": {
                                    "otp": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "9999"
                                    },
                                    "whatsapp_number": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "0502342345"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "GH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "'Success' : Beod has been successfully verified.'",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/generateOtp": {
            "post": {
                "tags": [
                    "Generate Otp"
                ],
                "summary": "Generate Otp",
                "description": "Generate OTP",
                "operationId": "generateOtp",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "whatsapp_number",
                                    "country_code"
                                ],
                                "properties": {
                                    "whatsapp_number": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "0502342345"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "GH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "\n    *               \n    *                   'Success': 'Otp has been sent to +233502785489'\n    *                   \n    *          ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "whatsapp_number": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "0502342345"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "GH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/waiting-list-registration": {
            "post": {
                "tags": [
                    "WaitingList"
                ],
                "summary": "Waiting list for users",
                "description": "Add user to waiting list",
                "operationId": "AddToWaitingList",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "whatsapp_number": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "0502342345"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "GH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": " 'Success' => 'You have successfully been added to the waitlist.' ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}